extract process control to process.go
